pyproject.toml
[project]
name = "cellophane"
version = "1.2.0.dev0"
description = "Library for wrappers integrating with SLIMS"
authors = [{name = "Erik Demitz-Helin",email = "erik.demitz-helin@gu.se"}]
readme = "README.md"
dependencies = [
    "click>=8.2.0,<9.0.0",
    "psutil>=5.9.4,<6.0.0",
    "rich-click>=1.8.9,<2.0.0",
    "humanfriendly>=10.0,<11.0",
    "jsonschema>=4.17.3,<5.0.0",
    "attrs>=23.1.0,<24.0.0",
    "ruamel-yaml>=0.17.32,<0.18.0",
    "mpire[dill]>=2.10.2",
    "gitpython>=3.1.31,<4.0.0",
    "questionary>=2.0.1,<3.0.0",
    "frozendict>=2.3.8,<3.0.0",
    "semver>=3.0.2,<4.0.0",
    "packaging>=24.0,<25.0",
    "xxhash>=3.4.1,<4.0.0",
    "pytest-profiling>=1.7.0,<2.0.0",
]
requires-python = ">=3.11,<4.0"

[tool]
[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true


[[tool.mypy.overrides]]
module = [
    "dill",
    "mpire",
    "mpire.exception",
    "mpire.async_result",
    "jsonschema",
]
ignore_errors = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "referencing.jsonschema"
ignore_errors = true


[tool.coverage.run]
omit = [
    "src/cellophane/testing/*",
    "src/cellophane/new_testing/*",
    "src/cellophane/__main__.py",
    "src/cellophane/__init__.py",
    "src/cellophane/src.py",
    "tests/*"
]
concurrency = [
    "multiprocessing",
    "thread"
]
branch = false

[tool.coverage.report]
omit = ["**/pytest-*/*"]



[project.scripts]
cellophane = "cellophane.dev:main"



[tool.ruff]
line-length = 88

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 88
exit-zero = true
fail-on = ["F","E","W"]

[tool.rye]
universal = true
managed = true
dev-dependencies = [
    "mypy>=0.991,<0.992",
    "bandit>=1.7.4,<2.0.0",
    "black>=22.12.0,<23.0.0",
    "flake8>=6.0.0,<7.0.0",
    "pytest>=8.3.3,<9.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "isort>=5.12.0,<6.0.0",
    "pytest-repeat>=0.9.1,<0.10.0",
    "pytest-clarity>=1.0.1,<2.0.0",
    "pytest-xdist>=3.3.1,<4.0.0",
    "types-jsonschema>=4.19.0.3,<5.0.0.0",
    "types-humanfriendly>=10.0.1.10,<11.0.0.0",
    "types-psutil>=5.9.5.16,<6.0.0.0",
    "pylint>=3.0.2,<4.0.0",
    "pytest-subprocess>=1.5.3",
]

[tool.hatch]
[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cellophane"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

